+++ snooty/gizaparser/test_steps.py
@@ -56,8 +56,9 @@ def test_step() -> None:
             "following command to import the\n</text><reference ",
             'refuri="https://www.mongodb.org/static/pgp/server-3.4.asc">',
             "<text>MongoDB public GPG Key</text>",
-            """</reference><target domain="std" name="label" target="mongodb-public-gpg-key" """,
+            """</reference><target domain="std" name="label" """,
             'refuri="https://www.mongodb.org/static/pgp/server-3.4.asc">',
+            """<target_identifier ids="['mongodb-public-gpg-key']"></target_identifier>"""
             "</target></paragraph></section></directive>",
             '<directive name="step">',
             "<section>",